+++ src/main/java/com/wine/to/up/catalog/service/domain/request/WinePositionTrueRequest.java
@@ -12,16 +12,16 @@ public class WinePositionTrueRequest {
     private WineTrueRequest wineTrueRequest;
 
     @JsonProperty("price")
-    private double price;
+    private float price;
 
     @JsonProperty("actual_price")
-    private double actual_price;
+    private float actual_price;
 
     @JsonProperty("link_to_wine")
     private String link_to_wine;
 
     @JsonProperty("volume")
-    private double volume;
+    private float volume;
 
     @JsonProperty("description")
     private String description;
